maintenance: restructured application files
