+++ src/auth/credentialsCrypto.ts
@@ -1,9 +1,9 @@
 import path from "path"
 import fsModule from "node:fs"
 import crypto from "node:crypto"
-import { platformConfigPath } from "./util"
+import { platformConfigPath } from "../util"
 import { type Credentials } from "./auth"
-import { key } from "./buildInfo"
+import { key } from "../buildInfo"
 
 /**
  * Handles cryptography for securely storing credentials in a file.
