+++ src/fs/autocompletion.ts
@@ -1,7 +1,7 @@
 import { CompleterResult } from "node:readline"
-import { Command, fsCommands, splitCommandSegments } from "./commands"
+import { Command, fsCommands, splitCommandSegments } from "../interface/commands"
 import FilenSDK from "@filen/sdk"
-import { CloudPath } from "./cloudPath"
+import { CloudPath } from "../cloudPath"
 import * as fs from "node:fs"
 import pathModule from "path"
