+++ src/fs/fs.ts
@@ -1,14 +1,14 @@
-import { err, out, outJson, prompt } from "./interface"
+import { err, out, outJson, prompt } from "../interface/interface"
 import FilenSDK from "@filen/sdk"
 import pathModule from "path"
-import { doNothing, formatBytes, formatTimestamp, hashFile } from "./util"
-import { CloudPath } from "./cloudPath"
+import { doNothing, formatBytes, formatTimestamp, hashFile } from "../util"
+import { CloudPath } from "../cloudPath"
 import cliProgress from "cli-progress"
 import * as fsModule from "node:fs"
 import { InterruptHandler } from "./interrupt"
 import open from "open"
-import { fsCommands } from "./commands"
-import { fsCommandsHelpPage } from "./helpPage"
+import { fsCommands } from "../interface/commands"
+import { fsCommandsHelpPage } from "../interface/helpPage"
 
 type CommandParameters = {
 	cloudWorkingPath: CloudPath
@@ -18,7 +18,8 @@ type CommandParameters = {
 }
 
 /**
- * Handles CLI commands related to cloud filesystem operations.
+ * Executes CLI commands related to cloud filesystem operations.
+ * @see FSInterface
  */
 export class FS {
 	private readonly filen: FilenSDK
