+++ src/fs/interrupt.ts
@@ -1,4 +1,4 @@
-import { out, readlineInterface } from "./interface"
+import { out, readlineInterface } from "../interface/interface"
 
 /**
  * Handles SIGINT signals.
